refine code

+++ server/errmsg/errmsg.go
@@ -14,12 +14,12 @@ type Error interface {
 }
 
 type err struct {
-	Code int         `json:"code"` // 业务编码
+	Code int64       `json:"code"` // 业务编码
 	Data interface{} `json:"data"` // 成功时返回的数据
 	Msg  string      `json:"msg"`  // 错误描述
 }
 
-func NewError(code int, msg string) Error {
+func NewError(code int64, msg string) Error {
 	return &err{
 		Code: code,
 		Data: nil,
@@ -35,7 +35,7 @@ func (e *err) WithData(data interface{}) Error {
 // ToString 返回 JSON 格式的错误详情
 func (e *err) ToString() string {
 	err := &struct {
-		Code int         `json:"code"`
+		Code int64       `json:"code"`
 		Data interface{} `json:"data"`
 		Msg  string      `json:"msg"`
 	}{
